+++ lib/src/page_model.dart
@@ -1,10 +1,25 @@
 import 'dart:ui';
 
-class PageModel {
-  final Color color;
-  final String imageAssetPath;
-  final String title;
-  final String body;
+import 'package:flutter/cupertino.dart';
 
-  PageModel(this.color, this.imageAssetPath, this.title, this.body);
+class PageModel {
+  Color color;
+  String imageAssetPath;
+  String title;
+  String body;
+  Widget child;
+  bool doAnimateChild;
+  bool doAnimateImage;
+
+  PageModel(
+      {@required this.color,
+      @required this.imageAssetPath,
+      @required this.title,
+      @required this.body,
+      @required this.doAnimateImage});
+
+  PageModel.withChild(
+      {@required this.child,
+      @required this.color,
+      @required this.doAnimateChild});
 }