+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## [1.0.5] - 289th August, 2019
+## [1.1.0-1.1.3] - 29th August, 2019
+* Option to pass complete widget with "PageModel.withChild"
+* Enable animation for child widget
+* Enable animation for page widget images
+
+## [1.0.5] - 29th August, 2019
 * Handle UI with many items
 
 ## [1.0.1-1.0.4] - 28th August, 2019